Make simulations reproducible using --seed
We make use of sets. Sets use object hashes and Python uses hash randomization to reduce the likelihood of denial of service attacks
[0]. However this means that between invocations of the Python interpreter, the hash of an object changes. So to get reproducible runs I've disabled hash randomization.

Users need to set the PYTHONHASHSEED environment variable to 0. There is a test checking for this.

Chose the current datetime in ISO format for the seed because it's easy to read and type out by hand.

[0]: https://docs.python.org/3/reference/datamodel.html#object.__hash__