Make simulations reproducible (#26)
* Default  --start-date is today at midnight

Trying to achieve reproducible runs and the changing time component is
stopping that.

* Specify ID for households rather than use id()

id() changes between runs. We can reproducibly generate an ID in BigQuery using farm_fingerprint.

* Make simulations reproducible using --seed

We make use of sets. Sets use object hashes and Python uses hash randomization to reduce the likelihood of denial of service attacks
[0]. However this means that between invocations of the Python interpreter, the hash of an object changes. So to get reproducible runs I've disabled hash randomization.

Users need to set the PYTHONHASHSEED environment variable to 0. There is a test checking for this.

Chose the current datetime in ISO format for the seed because it's easy to read and type out by hand.

[0]: https://docs.python.org/3/reference/datamodel.html#object.__hash__

* Rename function for clarity

* Update .env.template