Fix --heating-system-hassle-factor
This never worked in the first place because the value being passed to
`restrict_between_0_and_1` is a string, not a float, which has some
unusual behaviour:

```
>>> min('-0.1', 0) 0
>>> min('abcdf', 0) 0 ```

So `restrict_between_0_and_1` always returns 1.

This commit also adds [my suggestion][0] to raise an exception if the
value is less than 0 or greater than 1.

[0]: https://github.com/centrefornetzero/domestic-heating-abm/pull/11#discussion_r764005886