Type hint changes for stricter mypy checking
I almost got this to pass with `--strict` mode. Instead in
`pyproject.toml` I've specified every flag in `--strict` except for
`--disallow-any-generics`.

I'm not certain but I think our design is not type safe. The problem is
with the optional `model` argument to `Agent.make_decisions`, which is
of the argument type `Optional[AgentBasedModel]`. In subclasses of
`Agent` we want to specify a subclass of `AgentBasedModel`, but it's
[unsafe to override a method with a more specific argument type][0].

To make it type safe I think we'd have to make some more substantial
changes. I don't think there's sufficient benefit to justify the
effort.

[0]:
https://mypy.readthedocs.io/en/stable/common_issues.html#incompatible-overrides