Update scenario jobs (#120)
* Fix fuel prices

We are recalculating the fuel prices increase gas by 1.24p/kWh and
decreae electricity by 5.12p/kWh, based on our calculations of
environmental and social levies as a percentage of bills. Oil is
unchanged from default

* Remove `--price-gbp-per-kwh-oil`

Since it is the same as the default value, we can omit it from the args

* Add `annual_new_builds` arg to DomesticHeatingABM

Added:
* ANNUAL_NEW_BUILDS in constants.py
CCC / Element Energy projections of new builds, scaled to Eng/Wales
* Boolean flag to select this constant or not, and pass into
  ``DomesticHeatingABM`
* `new_builds_per_step` property to `DomesticHeatingABM` which scales
  the `annual_new_builds` to the step size, defaulting to zero
* `heat_pump_installation_capacity_new_builds` outputs
  `new_builds_per_step` post-2025, else 0
* `heat_pump_installation_capacity_existing_builds` which is the total
  capacity less the new build capacity

Updated:
* `has_heat_pump_installation_capacity` - the method called by the
  agents to check if they can install a heat pump. This will now take
  into account new builds as well, if applicable

* Include new builds in scenarios 4 and 5

Create a new scenario 4b which will model unlimited heat pump
installers. This will enable us to check what rate we should be training
heat pump installers at in order to meet new+existing build demand in
this maximum case

* Split out scenario 6 into two subscenarios

Consider a scenario with constrained and unconstrained heat pump
installers and compare in post-analysis

Co-authored-by: Tom Phillips <tom.phillips@centrefornetzero.org>